Finish splitting details cards
